layers: Dedicated bind_point.h
