+++ BUILD.gn
@@ -254,6 +254,7 @@ vvl_sources = [
   "layers/layer_options_validation.h",
   "layers/object_tracker/object_lifetime_validation.h",
   "layers/object_tracker/object_tracker_utils.cpp",
+  "layers/state_tracker/bind_point.h",
   "layers/state_tracker/buffer_state.cpp",
   "layers/state_tracker/buffer_state.h",
   "layers/state_tracker/cmd_buffer_state.cpp",
