+++ layers/core_checks/cc_pipeline_graphics.cpp
@@ -38,7 +38,6 @@
 #include "drawdispatch/drawdispatch_vuids.h"
 #include "chassis/chassis_modification_state.h"
 #include "state_tracker/image_state.h"
-#include "state_tracker/buffer_state.h"
 #include "state_tracker/descriptor_sets.h"
 #include "state_tracker/render_pass_state.h"
 #include "state_tracker/cmd_buffer_state.h"
